pyproject.toml
[project]
name = "labki-packs-tools"
version = "0.1.0"
description = "Tools for validating and visualizing Labki content packs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "jsonschema>=4.18",
]

[project.scripts]
labki-validate = "labki_packs_tools.validation.cli:main"
labki-graph = "tools.graph_repo:main"

[project.optional-dependencies]
tests = [
    "pytest>=8.4.2",
]
dev = [
    "labki-packs-tools[tests]",
    "ruff>=0.13.2",
    "black>=25.9.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.scripts]
test = "pytest"
lint.composite = [
    "ruff check",
    "black . --diff",
]
format.composite = [
    "black .",
    "ruff check --fix",
]


[tool.ruff]
target-version = "py310"
include = ["src/labki_packs_tools/**/*.py", "tests/**/*.py", "pyproject.toml"]
exclude = ["docs"]
line-length = 100

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # annotations
    "ANN",
    # no debugger
    "T10",
    ## ----------
    # pydocstyle
    # indentation
    "D207", "D208",
    # whitespace
    "D210", "D211",
    # emptiness
    "D419",
    # snake case function name
    "N802",
]
ignore = [
    "ANN401", "ANN204",
    # explicit strict arg for zip
    "B905",
    # builtin type annotations
    "UP006", "UP035",
    # | for Union types (only supported >=3.10
    "UP007",
    # using Optional[] is fine
    "UP045",
    # docstrings for __init__
    "D107",
    # Sometimes expanding if/else is simpler
    "SIM114",
]

fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"tests/**/test_*" = ["ANN001", "ANN201", "F841"]

[tool.black]
target-version = ['py311', 'py312', 'py313']
include = "(?:tests|src)/.*\\.py$"
line-length = 100

[tool.pytest.ini_options]
markers = """
  packaging: meta-tests that confirm equivalent behavior in editable/git installs, sdist, and wheels
"""
